[Enh]: Added nix-homemanager bootstrap.
Mostly WIP because full installation is not tested (it needs some channels checks mentioned in `install-homenager.sh`) and whole concept of integrating nix + ansible is raw.

Nevertheless, as for me, nix is better for managing current machine. It's not possible on non-NixOs system to manage systemd and other directories like /etc though.
That's where ansible comes!

Moreover, it's cool to use ansible to bootstrap nix on machines (name of the repo after all) because official instructions are quite complicated.