[Fix](Filesystem): Changed zfs /nix mountpoint.
Manual mount --bind wasn't working on system boot. It seems as problems with
timings because manually mounting from config /etc/fstab works fine.

So just moved mountpoint because it's not that important now. Maybe
in the future, when I will have several stores in one point for sharing things,
it will be necessary to make adjustments.